VSTelemetry Api changes.
We were originally setting properties with a `IsOptOutFriendly` option, so that sdk version information
would still be sent even if the user opted out. We can’t do that, in that we can’t set individual properties
on a case by case basis. We can send an event that is ‘IsOptOutFriendly’ so we will do that instead.