Sort MobileProvisionIndex records by CreationDate
This patch doesn't "fix" existing indexes to be sorted by
CreationDate, but if/when any new records are added, the
index will get sorted.

The advantage of sorting by CreationDate allows us to avoid
calling MobileProvision.Load() for anything but the *latest*
version of a provisioning profile when 'unique' is 'true'.

Since 'unique' is alays 'true' in the iOS Bundle Signing
Project Options, this could theoretically improve performance
if the user has a bunch of duplicate provisioning profiles
(which can happen if devices and/or developers have been
added/removed from a provisioning profile, the entitlements
have changed, or if the provisioning profile has been renewed).

*May* fix https://bugzilla.xamarin.com/show_bug.cgi?id=54150