Massively optimized MobileProvisionIndex by using custom data structures
The problem with using a PList is that when you have massive
PArrays of dev certs, it causes saving the plist to be unusably
slow due to all of the linear lookups to get the indexes.

There was also overhead of PDictionary over the use of direct
int/string/etc properties.

Fixes perf found in https://bugzilla.xamarin.com/show_bug.cgi?id=58678