Adjust SDK validation to allow tools/bin and not allow usr/bin.
Adjust SDK validation to find mtouch in a tools/bin directory (which is where
mtouch resides when shipped in a nuget package), and not look in usr/bin
(mtouch hasn't been there in a long, long time, and then it was very briefly).

Simplify the validation logic to output mtouch's actual path instead of a
value indicating the subdirectory where it was found.

Additionally compute the bin directory to be mtouch's directory, and the lib
directory a sibling of the bin directory.